ptc/account_exporter/requirements.txt
mysql-connector==2.2.9
requests
apscheduler==3.10.4